Managing technical debt
Replaced Tabs with Spaces to conform with PEP8
Separated Frame and Signal objects into their own modules
Lark files have been moved into their own folder
